Revert the previous revision.
This would result in two errors.

First an fsck is not in progress just because fsck is running; this is
precisely why this flag is different - we need to wait until we get the
first progress indicator on the pipe.

Secondly the change from > to >= would mean that the currently displayed
message would change.  The user might have just pressed "S" to skip one
filesystem, but mountall now decides to skip a completely different one.

Bad.